Fix deletion of temporary platform files:
Deletion of platform files was done by the run method in
JubeBenchmarkingAPI class, this behavior causes bugs when the run
method is executed several times. This was solved by allowing the
JubeXMLParser to handle the deletion of platform file when the object
is not longer referenced (__del__ method).

Cosmetic changes were added.