adds Publisher class main functionality
Provides major changes in modules publish and git. Adds new module
standard_vcs. This commit changes earlier Publisher class signature
call and for that reason the CampaignManager class had to be updated
New environement variables for the Publisher class were aded and they
are managed by the ubench info command from the UbenchConfig class
(which is also part of this commit).