Revert transposition for ubench results file
The "transpose" parameter in Jube allows the output to be more readable when
you have many more columns than lines, but it breaks the generation of the
unclebench "bench_results.yaml" file ("results_bench: failed").
One can always edit this parameter in the "configuration.xml" file of the bench
and regenerate the results, but you may have to create a pseudo-platform first.
A nicer solution is to revert the transposition for "result/ubench_results.dat"
only, which is used for "bench_results.yaml", without modifying other files
or outputs.