Remove spaces when parsing use steps
The <use> steps "ubench_config" and "ubench_files" are added when missing
for multisources functionnalities.
But when the xml file has spaces (e.g. "<use> ubench_config </use>"), the code
thinks the step is missing, adds it, and the resulting  xml is rejected by Jube
because of a duplicate key.